Changed back bit offset positions to long
RB=642915
G=pinot-dev-reviewers
R=kgopalak,jfim,ssubrama,mshrivas
A=jfim

+++ pinot-core/src/main/java/com/linkedin/pinot/core/io/reader/impl/FixedBitSingleValueMultiColReader.java
@@ -226,8 +226,8 @@ public class FixedBitSingleValueMultiColReader implements Closeable {
    * @param col
    * @return
    */
-  private int computeBitOffset(int row, int col) {
-    final int offset = row * rowSizeInBits + colBitOffSets[col];
+  private long computeBitOffset(int row, int col) {
+    final long offset = ( (long)row * rowSizeInBits ) + colBitOffSets[col];
     return offset;
   }
